Add files via upload
In this script, the iptables command is used to modify the system's firewall rules. The -A INPUT -p tcp --destination-port $i -j DROP part of the command adds a new rule to the INPUT chain that drops all incoming TCP packets with a destination port of $i.

Please note that this script must be run with root privileges, as changing firewall rules typically requires root access. Also note that these changes are not persistent: they will be lost if the system is rebooted. If you want to make these changes persistent, you would need to save the iptables rules to a file and restore them at boot time, or use a tool like iptables-persistent.

Furthermore, it's important to understand that blocking these ports could have implications for your system and network. Ports 1 through 5 are well-known ports used by various network protocols, and blocking them could disrupt these services. Before making any changes to your firewall settings, you should ensure you understand the potential consequences.

Also, this script is specifically for Unix-like systems with iptables installed. If you are using a different type of system or firewall tool, the specific commands would be different.