run as admin
In this script, the New-NetFirewallRule command is used to create a new inbound rule that blocks TCP traffic on the specified port.

You would need to run this script with administrative privileges. To do this, you can right-click the PowerShell icon and select "Run as Administrator".

Please note that, as with the iptables example above, blocking these ports could disrupt network services that rely on them. You should ensure you understand the potential impact of these changes before implementing them.

This script is designed for modern versions of Windows that include the New-NetFirewallRule PowerShell cmdlet. If you are using an older version of Windows, the specific commands may be different.